pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cosmicsnip"
version = "1.0.0"
description = "A polished screenshot snipping tool for COSMIC Desktop / Wayland"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "itssoup" }]
keywords = ["screenshot", "snipping-tool", "cosmic", "wayland", "gtk4"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Wayland",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Graphics :: Capture :: Screen Capture",
]
dependencies = [
    "PyGObject>=3.42",
    "Pillow>=9.0",
    "dbus-python>=1.3",
]

[project.urls]
Homepage = "https://github.com/taran3030/cosmicsnip"
Issues = "https://github.com/taran3030/cosmicsnip/issues"

[project.scripts]
cosmicsnip = "cosmicsnip.app:main"

[project.gui-scripts]
cosmicsnip-gui = "cosmicsnip.app:main"

[tool.setuptools.packages.find]
include = ["cosmicsnip*"]
